web: Refactor UI code
 * Created directory for UI
 * Moved rendering code to js/ui/render.js
 * Created initial js/ui/util.js for UI utilities

Notice I did not use modules. I was thinking of doing this but I am
avoiding it for now as it is much easier to just have everything
declared as global at this time. For onclick methods this makes sense;
but for code (such as rendering) I think it would be better off as a
module that gets imported. This is a task for another day.