web: fix anchor detection
This fixes an issue where text was being hyperlinked when the "http"
pattern was detected in a character block. This looks that there is
whitespace or such to correctly assume it's a link. This may not be the
best implementation, but it seems to resolve issues for me so far. Unit
tests would be good.