Adding Charlie Fish to team page
I **think** at this point I’ve made enough Damus contributions to be considered part of the team maybe?

If not, no hard feelings.

One possibly controversial thing is that I’m pointing to my own server for the profile picture. This will reduce the number of places that I have to update my profile picture if I ever do in the future. So a lot easier for me, and one less thing for me to possibly forget.

I checked and damus doesn’t have a Content-Security-Policy so this should be fine technically.

Signed-off-by: Charlie Fish <contact@charlie.fish>
Signed-off-by: William Casarin <jb55@jb55.com>