Improve error handling UI/UX on LN checkout
This commit adds a mechanism to display errors nicely to the user.

All (or most) functions related to the checkout process were wrapped in
try-catch statements that display error nicely with support contact info
and a reference number.

Detailed error messages were written containing tips for the users.

Testing
--------

Test 1: Went through the entire LN flow locally, and it worked without errors
Test 2: Went through the entire LN flow, turning the server on and off and the wi-fi on/off in each step. Saw the nice messages appearing.

Signed-off-by: Daniel D’Aquino <daniel@daquino.me>
Link: 20240131021044.86169-3-daniel@daquino.me
Signed-off-by: William Casarin <jb55@jb55.com>