Let LN node communication have a few tries before displaying errors.
On staging it was observed that occasionally LN node communication fails
once or twice before succeeding. This commit allows such functions to
try a few times before showing the user that there was an error.

This is a preventative measure to avoid sensitive false positives.

Testing: Ran LN flow locally to verify no error messages are displayed.
Signed-off-by: Daniel D’Aquino <daniel@daquino.me>
Link: 20240131021044.86169-4-daniel@daquino.me
Signed-off-by: William Casarin <jb55@jb55.com>