Improve checkout refresh stability
This commit improves the stability when refreshing the checkout page.

More specifically, it avoids an issue where the person has selected
Damus iOS as their selected auth method, but then refreshes the page
after verifying their npub. Previously the final step would not be aware
of their auth method, so it would try to proceed on the web, but a lack
of session token would cause issues on the welcome step.

This commit addresses this by checking the login status at the last
step. If the user selected the OTP auth method, they will necessarily be
logged in already. If the user is not logged in, they cannot proceed
through the website.

Signed-off-by: Daniel D’Aquino <daniel@daquino.me>