Fix checkout and login for npubs devoid of a profile
This commit fixes the checkout and login flows for pubkeys without a
profile.

The recent changes in the checkout and login flows caused a regression
where a pubkey without any profiles would simply not show up, and it
would make it impossible for the user to proceed.

This commit fixes it by displaying the nostr profile and continue button
when the profile is null (meaning that it was loaded but no profiles
were found), showing a robohash icon and an abbreviated npub.

Signed-off-by: Daniel D’Aquino <daniel@daquino.me>